refactor: 알림 이벤트 발행 클래스명 및 메서드 명 수정
